plant_disease_detection/requirements.txt
# Core ML and DL
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
scikit-learn==1.3.2

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# API and Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# UI
streamlit==1.28.2
plotly==5.18.0

# Data handling
pandas==2.1.3

# Load testing
locust==2.19.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
matplotlib==3.8.2
seaborn==0.13.0